Extract site cell into its own component
