+++ packages/domains-table/src/domains-table/__tests__/domains-table-row.tsx
@@ -3,7 +3,7 @@
  */
 import { screen, waitFor } from '@testing-library/react';
 import React from 'react';
-import { renderWithProvider, testDomain, testPartialDomain, testSite } from '../../test-utils';
+import { renderWithProvider, testDomain, testPartialDomain } from '../../test-utils';
 import { DomainsTableRow } from '../domains-table-row';
 
 const noop = jest.fn();
@@ -219,90 +219,4 @@ test( 'transfer domains link to the transfer management interface', async () => 
 		'href',
 		'/domains/manage/example.com/transfer/in/example.com'
 	);
-} );
-
-describe( 'columns', () => {
-	test( 'when a site is associated with a domain, display its name', async () => {
-		const [ primaryPartial, primaryFull ] = testDomain( {
-			domain: 'primary-domain.blog',
-			blog_id: 123,
-			primary_domain: true,
-		} );
-
-		const fetchSiteDomains = jest.fn().mockImplementation( () =>
-			Promise.resolve( {
-				domains: [ primaryFull ],
-			} )
-		);
-
-		const fetchSite = jest
-			.fn()
-			.mockImplementation( () =>
-				Promise.resolve( testSite( { ID: 123, name: 'Primary Domain Blog' } ) )
-			);
-
-		render(
-			<DomainsTableRow
-				domain={ primaryPartial }
-				isAllSitesView
-				fetchSite={ fetchSite }
-				fetchSiteDomains={ fetchSiteDomains }
-				isSelected={ false }
-				onSelect={ noop }
-			/>
-		);
-
-		await waitFor( () =>
-			expect( screen.queryByText( 'Primary Domain Blog' ) ).toBeInTheDocument()
-		);
-	} );
-
-	test( 'when a site is not associated with a domain, display site linking ctas', async () => {
-		const [ primaryPartial, primaryFull ] = testDomain( {
-			domain: 'primary-domain.blog',
-			blog_id: 123,
-			primary_domain: true,
-			current_user_can_create_site_from_domain_only: true,
-		} );
-
-		const fetchSiteDomains = jest.fn().mockImplementation( () =>
-			Promise.resolve( {
-				domains: [ primaryFull ],
-			} )
-		);
-
-		const fetchSite = jest
-			.fn()
-			.mockImplementation( () =>
-				Promise.resolve( testSite( { URL: 'https://primary-domain.blog', ID: 123 } ) )
-			);
-
-		render(
-			<DomainsTableRow
-				domain={ primaryPartial }
-				isAllSitesView
-				fetchSite={ fetchSite }
-				fetchSiteDomains={ fetchSiteDomains }
-				isSelected={ false }
-				onSelect={ noop }
-			/>
-		);
-
-		await waitFor( () => {
-			const createLink = screen.queryByText( 'Create' );
-			const connectLink = screen.queryByText( 'connect' );
-
-			expect( createLink ).toBeInTheDocument();
-			expect( createLink ).toHaveAttribute(
-				'href',
-				'/start/site-selected/?siteSlug=primary-domain.blog&siteId=123'
-			);
-
-			expect( connectLink ).toBeInTheDocument();
-			expect( connectLink ).toHaveAttribute(
-				'href',
-				'/domains/manage/all/primary-domain.blog/transfer/other-site/primary-domain.blog'
-			);
-		} );
-	} );
 } );
